--- Clock.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.6.33815.320
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Clock", "Clock\Clock.vcxproj", "{4BCBD133-BCA0-4000-9A57-FD4793928AFF}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4BCBD133-BCA0-4000-9A57-FD4793928AFF}.Debug|x64.ActiveCfg = Debug|x64
-		{4BCBD133-BCA0-4000-9A57-FD4793928AFF}.Debug|x64.Build.0 = Debug|x64
-		{4BCBD133-BCA0-4000-9A57-FD4793928AFF}.Debug|x86.ActiveCfg = Debug|Win32
-		{4BCBD133-BCA0-4000-9A57-FD4793928AFF}.Debug|x86.Build.0 = Debug|Win32
-		{4BCBD133-BCA0-4000-9A57-FD4793928AFF}.Release|x64.ActiveCfg = Release|x64
-		{4BCBD133-BCA0-4000-9A57-FD4793928AFF}.Release|x64.Build.0 = Release|x64
-		{4BCBD133-BCA0-4000-9A57-FD4793928AFF}.Release|x86.ActiveCfg = Release|Win32
-		{4BCBD133-BCA0-4000-9A57-FD4793928AFF}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {3C5E31C9-4A1A-4FE5-BC9B-083A3A176660}
-	EndGlobalSection
-EndGlobal
