--- Clock/x64/Debug/Clock.log
@@ -1,2 +1,0 @@
-﻿  Clock.cpp
-  Clock.vcxproj -> C:\Users\Maxance\Desktop\plock_win\x64\Debug\Clock.exe
