--- Clock/x64/Debug/Clock.tlog/Cl.items.tlog
@@ -1,1 +1,0 @@
-C:\Users\Maxance\Desktop\plock_win\Clock\Clock.cpp;C:\Users\Maxance\Desktop\plock_win\Clock\x64\Debug\Clock.obj
