--- Clock/x64/Debug/Clock.tlog/Clock.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.39.33519:TargetPlatformVersion=10.0.22621.0:
-Debug|x64|C:\Users\Maxance\Desktop\plock_win\|
