--- Clock/x64/Debug/Clock.tlog/link.secondary.1.tlog
@@ -1,2 +1,0 @@
-^C:\USERS\MAXANCE\DESKTOP\PLOCK_WIN\CLOCK\X64\DEBUG\CLOCK.OBJ
-C:\Users\Maxance\Desktop\plock_win\Clock\x64\Debug\Clock.ilk
