+++ plock_win.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34902.65
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "plock_win", "plock_win\plock_win.vcxproj", "{ED0CF534-E1E5-44B8-9AB1-0073D5E493BC}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{ED0CF534-E1E5-44B8-9AB1-0073D5E493BC}.Debug|x64.ActiveCfg = Debug|x64
+		{ED0CF534-E1E5-44B8-9AB1-0073D5E493BC}.Debug|x64.Build.0 = Debug|x64
+		{ED0CF534-E1E5-44B8-9AB1-0073D5E493BC}.Debug|x86.ActiveCfg = Debug|Win32
+		{ED0CF534-E1E5-44B8-9AB1-0073D5E493BC}.Debug|x86.Build.0 = Debug|Win32
+		{ED0CF534-E1E5-44B8-9AB1-0073D5E493BC}.Release|x64.ActiveCfg = Release|x64
+		{ED0CF534-E1E5-44B8-9AB1-0073D5E493BC}.Release|x64.Build.0 = Release|x64
+		{ED0CF534-E1E5-44B8-9AB1-0073D5E493BC}.Release|x86.ActiveCfg = Release|Win32
+		{ED0CF534-E1E5-44B8-9AB1-0073D5E493BC}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {53FA6F7E-EF54-4F59-BBB6-3FF01674055C}
+	EndGlobalSection
+EndGlobal
