+++ plock_win/x64/Debug/plock_win.log
@@ -1,0 +1,2 @@
+﻿  plock_win.cpp
+  plock_win.vcxproj -> C:\Users\Maxance\Desktop\plock_win\plock_win\x64\Debug\plock_win.exe
