+++ plock_win/x64/Debug/plock_win.tlog/Cl.items.tlog
@@ -1,0 +1,1 @@
+C:\Users\Maxance\Desktop\plock_win\plock_win\plock_win\plock_win.cpp;C:\Users\Maxance\Desktop\plock_win\plock_win\plock_win\x64\Debug\plock_win.obj
