+++ plock_win/x64/Debug/plock_win.tlog/link.secondary.1.tlog
@@ -1,0 +1,2 @@
+^C:\USERS\MAXANCE\DESKTOP\PLOCK_WIN\PLOCK_WIN\PLOCK_WIN\X64\DEBUG\PLOCK_WIN.OBJ
+C:\Users\Maxance\Desktop\plock_win\plock_win\plock_win\x64\Debug\plock_win.ilk
