+++ plock_win/x64/Debug/plock_win.tlog/plock_win.lastbuildstate
@@ -1,0 +1,2 @@
+PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.39.33519:TargetPlatformVersion=10.0.22621.0:
+Debug|x64|C:\Users\Maxance\Desktop\plock_win\plock_win\|
